added support for generating html straight out of makeinfo (you need v4.0 for this, so it's commented out at the moment. Unlike texi2html, the html output from makeinfo is a single monolithic file; not too great..)
Bug-fix (Well, I think so..) If you do a "make" and then "su root" and
"make install" and then return to non-root, the install would leave 2 files
in the source tree that were owned by root - very impolite. The files were:
-- gforth.fi (I solve this problem by deleting gforth.fi after an
install. I think gforth.fi after an install refers to the install directory
so this is a good thing to do anyway)
-- prim.TAGS (because "make all" doesn't generate prim.TAGS, but
"make install" does, therefore it ends up being owned by root. I couldn't
see a clean way to fix this in the Makefile. I'd use "chmod --reference"
to just fix up the protection but this option is GNU-specific. My backup
solution was to add TAGS to the list in the doc: target, which seemed
reasonable/consistent).